+++ geotrek/flatpages/forms.py
@@ -1,6 +1,6 @@
 from django.conf import settings
 from django.core import validators
-from tinymce.widgets import TinyMCE
+from geotrek.flatpages.widgets import FlatPageTinyMCE
 
 from treebeard.forms import MoveNodeForm
 
@@ -26,7 +26,7 @@ class FlatPageForm(MoveNodeForm):
         # are passed along as they contain modeltranslation CSS classes.
         for fieldname, formfield in self.fields.items():
             if fieldname.startswith('content_'):
-                self.fields[fieldname].widget = TinyMCE(attrs=self.fields[fieldname].widget.attrs)
+                self.fields[fieldname].widget = FlatPageTinyMCE(attrs=self.fields[fieldname].widget.attrs)
 
         if self.instance.pk:
             page = Attachment.objects.filter(