+++ docs/changelog.rst
@@ -8,6 +8,7 @@ CHANGELOG
 **Features**
 
 - Add new menu headers and flat pages with tree hierarchies
+- Add styling and embedding tools for FlatPages content (refs #3921, #3922, #4019)
 
 **Bug fixes**